requirements.txt
django
djangorestframework
drf-yasg
django-filter
djangorestframework-simplejwt
django-cors-headers
Pillow
stripe